refactor(web): split the three modal and settings size-ratchet functions
Removes three max-lines-per-function suppressions:

- DataStorageSettings rendered the request-log and app-log purge flows
  as two copies of the same button, range select, confirm and cancel.
  One PurgeLogsControl now owns the confirm and selection state and is
  rendered twice with its labels and mutation; the parent keeps the
  toasts and cache invalidation.
- UserModal's form state, validation and four mutations moved into
  useUserForm, returned as one bag so the component is markup only.
- KeyDetailModal's edit state, owner-cap derivation and two mutations
  moved into useKeyEdit the same way.

No behaviour change; the existing modal and settings tests render every
moved branch.
